+++ protocols/kad/src/record/store/memory.rs
@@ -44,7 +44,7 @@ pub struct MemoryStore {
 }
 
 /// Configuration for a `MemoryStore`.
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct MemoryStoreConfig {
     /// The maximum number of records.
     pub max_records: usize,