+++ swarm/src/connection/listeners.rs
@@ -25,59 +25,22 @@ use crate::{
     Multiaddr, Transport,
 };
 use futures::{prelude::*, task::Context, task::Poll};
+use libp2p_core::connection::ListenerId;
 use log::debug;
 use smallvec::SmallVec;
 use std::{collections::VecDeque, fmt, mem, pin::Pin};
 
 /// Implementation of `futures::Stream` that allows listening on multiaddresses.
 ///
-/// To start using a `ListenersStream`, create one with `new` by passing an implementation of
-/// `Transport`. This `Transport` will be used to start listening, therefore you want to pass
-/// a `Transport` that supports the protocols you wish you listen on.
+/// To start using a [`ListenersStream`], create one with [`ListenersStream::new`] by passing an
+/// implementation of [`Transport`]. This [`Transport`] will be used to start listening, therefore
+/// you want to pass a [`Transport`] that supports the protocols you wish you listen on.
 ///
-/// Then, call `ListenerStream::listen_on` for all addresses you want to start listening on.
+/// Then, call [`ListenersStream::listen_on`] for all addresses you want to start listening on.
 ///
-/// The `ListenersStream` never ends and never produces errors. If a listener errors or closes,
-/// an event is generated on the stream and the listener is then dropped, but the `ListenersStream`
+/// The [`ListenersStream`] never ends and never produces errors. If a listener errors or closes, an
+/// event is generated on the stream and the listener is then dropped, but the [`ListenersStream`]
 /// itself continues.
-///
-/// # Example
-///
-/// ```no_run
-/// use futures::prelude::*;
-/// use libp2p_core::connection::{ListenersEvent, ListenersStream};
-///
-/// let mut listeners = ListenersStream::new(libp2p_tcp::TcpConfig::new());
-///
-/// // Ask the `listeners` to start listening on the given multiaddress.
-/// listeners.listen_on("/ip4/0.0.0.0/tcp/0".parse().unwrap()).unwrap();
-///
-/// // The `listeners` will now generate events when polled.
-/// futures::executor::block_on(async move {
-///     while let Some(event) = listeners.next().await {
-///         match event {
-///             ListenersEvent::NewAddress { listener_id, listen_addr } => {
-///                 println!("Listener {:?} is listening at address {}", listener_id, listen_addr);
-///             },
-///             ListenersEvent::AddressExpired { listener_id, listen_addr } => {
-///                 println!("Listener {:?} is no longer listening at address {}", listener_id, listen_addr);
-///             },
-///             ListenersEvent::Closed { listener_id, .. } => {
-///                 println!("Listener {:?} has been closed", listener_id);
-///             },
-///             ListenersEvent::Error { listener_id, error } => {
-///                 println!("Listener {:?} has experienced an error: {}", listener_id, error);
-///             },
-///             ListenersEvent::Incoming { listener_id, upgrade, local_addr, .. } => {
-///                 println!("Listener {:?} has a new connection on {}", listener_id, local_addr);
-///                 // We don't do anything with the newly-opened connection, but in a real-life
-///                 // program you probably want to use it!
-///                 drop(upgrade);
-///             },
-///         }
-///     }
-/// })
-/// ```
 pub struct ListenersStream<TTrans>
 where
     TTrans: Transport,
@@ -93,13 +56,6 @@ where
     /// Pending listeners events to return from [`ListenersStream::poll`].
     pending_events: VecDeque<ListenersEvent<TTrans>>,
 }
-
-/// The ID of a single listener.
-///
-/// It is part of most [`ListenersEvent`]s and can be used to remove
-/// individual listeners from the [`ListenersStream`].
-#[derive(Copy, Clone, Debug, PartialEq, Eq, Hash, PartialOrd, Ord)]
-pub struct ListenerId(u64);
 
 /// A single active listener.
 #[pin_project::pin_project]
@@ -178,18 +134,7 @@ where
         ListenersStream {
             transport,
             listeners: VecDeque::new(),
-            next_id: ListenerId(1),
-            pending_events: VecDeque::new(),
-        }
-    }
-
-    /// Same as `new`, but pre-allocates enough memory for the given number of
-    /// simultaneous listeners.
-    pub fn with_capacity(transport: TTrans, capacity: usize) -> Self {
-        ListenersStream {
-            transport,
-            listeners: VecDeque::with_capacity(capacity),
-            next_id: ListenerId(1),
+            next_id: ListenerId::new(1),
             pending_events: VecDeque::new(),
         }
     }
@@ -211,7 +156,7 @@ where
             addresses: SmallVec::new(),
         }));
         let id = self.next_id;
-        self.next_id = ListenerId(self.next_id.0 + 1);
+        self.next_id = self.next_id + 1;
         Ok(id)
     }