+++ swarm/src/connection/pool/task.rs
@@ -24,24 +24,21 @@
 use super::concurrent_dial::ConcurrentDial;
 use crate::{
     connection::{
-        self,
-        handler::{THandlerError, THandlerInEvent, THandlerOutEvent},
-        ConnectionError, ConnectionHandler, ConnectionId, IntoConnectionHandler,
-        PendingInboundConnectionError, PendingOutboundConnectionError, Substream,
+        self, ConnectionError, PendingInboundConnectionError, PendingOutboundConnectionError,
     },
-    muxing::StreamMuxer,
     transport::{Transport, TransportError},
-    Multiaddr, PeerId,
+    Multiaddr, PeerId, ProtocolsHandler,
 };
 use futures::{
     channel::{mpsc, oneshot},
     future::{poll_fn, Either, Future},
     SinkExt, StreamExt,
 };
+use libp2p_core::connection::ConnectionId;
 use std::pin::Pin;
 use void::Void;
 
-/// Commands that can be sent to a task.
+/// Commands that can be sent to a task driving an established connection.
 #[derive(Debug)]
 pub enum Command<T> {
     /// Notify the connection handler of an event.
@@ -75,7 +72,7 @@ where
 }
 
 #[derive(Debug)]
-pub enum EstablishedConnectionEvent<THandler: IntoConnectionHandler> {
+pub enum EstablishedConnectionEvent<THandler: ProtocolsHandler> {
     /// A node we are connected to has changed its address.
     AddressChange {
         id: ConnectionId,
@@ -86,7 +83,7 @@ pub enum EstablishedConnectionEvent<THandler: IntoConnectionHandler> {
     Notify {
         id: ConnectionId,
         peer_id: PeerId,
-        event: THandlerOutEvent<THandler>,
+        event: THandler::OutEvent,
     },
     /// A connection closed, possibly due to an error.
     ///
@@ -95,20 +92,20 @@ pub enum EstablishedConnectionEvent<THandler: IntoConnectionHandler> {
     Closed {
         id: ConnectionId,
         peer_id: PeerId,
-        error: Option<ConnectionError<THandlerError<THandler>>>,
-        handler: THandler::Handler,
+        error: Option<ConnectionError<THandler::Error>>,
+        handler: THandler,
     },
 }
 
 pub async fn new_for_pending_outgoing_connection<TTrans>(
     connection_id: ConnectionId,
     dial: ConcurrentDial<TTrans>,
-    drop_receiver: oneshot::Receiver<Void>,
+    abort_receiver: oneshot::Receiver<Void>,
     mut events: mpsc::Sender<PendingConnectionEvent<TTrans>>,
 ) where
     TTrans: Transport,
 {
-    match futures::future::select(drop_receiver, Box::pin(dial)).await {
+    match futures::future::select(abort_receiver, Box::pin(dial)).await {
         Either::Left((Err(oneshot::Canceled), _)) => {
             let _ = events
                 .send(PendingConnectionEvent::PendingFailed {
@@ -141,13 +138,13 @@ pub async fn new_for_pending_outgoing_connection<TTrans>(
 pub async fn new_for_pending_incoming_connection<TFut, TTrans>(
     connection_id: ConnectionId,
     future: TFut,
-    drop_receiver: oneshot::Receiver<Void>,
+    abort_receiver: oneshot::Receiver<Void>,
     mut events: mpsc::Sender<PendingConnectionEvent<TTrans>>,
 ) where
     TTrans: Transport,
     TFut: Future<Output = Result<TTrans::Output, TTrans::Error>> + Send + 'static,
 {
-    match futures::future::select(drop_receiver, Box::pin(future)).await {
+    match futures::future::select(abort_receiver, Box::pin(future)).await {
         Either::Left((Err(oneshot::Canceled), _)) => {
             let _ = events
                 .send(PendingConnectionEvent::PendingFailed {
@@ -179,16 +176,14 @@ pub async fn new_for_pending_incoming_connection<TFut, TTrans>(
     }
 }
 
-pub async fn new_for_established_connection<TMuxer, THandler>(
+pub async fn new_for_established_connection<THandler>(
     connection_id: ConnectionId,
     peer_id: PeerId,
-    mut connection: crate::connection::Connection<TMuxer, THandler::Handler>,
-    mut command_receiver: mpsc::Receiver<Command<THandlerInEvent<THandler>>>,
+    mut connection: crate::connection::Connection<THandler>,
+    mut command_receiver: mpsc::Receiver<Command<THandler::InEvent>>,
     mut events: mpsc::Sender<EstablishedConnectionEvent<THandler>>,
 ) where
-    TMuxer: StreamMuxer,
-    THandler: IntoConnectionHandler,
-    THandler::Handler: ConnectionHandler<Substream = Substream<TMuxer>>,
+    THandler: ProtocolsHandler,
 {
     loop {
         match futures::future::select(